backend/app/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
chromadb
sentence-transformers
openai
httpx
pydantic
python-dotenv
requests